changed: put settings related videodatabase methods in separate file
