+++ xbmc/video/database/VideoDatabaseSettings.cpp
@@ -1,0 +1,561 @@
+/*
+ *  Copyright (C) 2016-2018 Team Kodi
+ *  This file is part of Kodi - https://kodi.tv
+ *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSES/README.md for more information.
+ */
+
+#include "VideoDatabase.h"
+
+#include <string>
+
+#include "addons/AddonManager.h"
+#include "dbwrappers/dataset.h"
+#include "cores/VideoSettings.h"
+#include "dialogs/GUIDialogProgress.h"
+#include "FileItem.h"
+#include "filesystem/MultiPathDirectory.h"
+#include "utils/log.h"
+#include "utils/URIUtils.h"
+#include "video/VideoInfoScanner.h"
+
+using namespace ADDON;
+using namespace VIDEO;
+using namespace XFILE;
+
+bool CVideoDatabase::GetVideoSettings(const CFileItem &item, CVideoSettings &settings)
+{
+  return GetVideoSettings(GetFileId(item), settings);
+}
+
+/// \brief GetVideoSettings() obtains any saved video settings for the current file.
+/// \retval Returns true if the settings exist, false otherwise.
+bool CVideoDatabase::GetVideoSettings(const std::string &filePath, CVideoSettings &settings)
+{
+  return GetVideoSettings(GetFileId(filePath), settings);
+}
+
+bool CVideoDatabase::GetVideoSettings(int idFile, CVideoSettings &settings)
+{
+  try
+  {
+    if (idFile < 0) return false;
+    if (NULL == m_pDB.get()) return false;
+    if (NULL == m_pDS.get()) return false;
+
+    std::string strSQL=PrepareSQL("select * from settings where settings.idFile = '%i'", idFile);
+    m_pDS->query( strSQL );
+
+    if (m_pDS->num_rows() > 0)
+    { // get the video settings info
+      settings.m_AudioDelay = m_pDS->fv("AudioDelay").get_asFloat();
+      settings.m_AudioStream = m_pDS->fv("AudioStream").get_asInt();
+      settings.m_Brightness = m_pDS->fv("Brightness").get_asFloat();
+      settings.m_Contrast = m_pDS->fv("Contrast").get_asFloat();
+      settings.m_CustomPixelRatio = m_pDS->fv("PixelRatio").get_asFloat();
+      settings.m_CustomNonLinStretch = m_pDS->fv("NonLinStretch").get_asBool();
+      settings.m_NoiseReduction = m_pDS->fv("NoiseReduction").get_asFloat();
+      settings.m_PostProcess = m_pDS->fv("PostProcess").get_asBool();
+      settings.m_Sharpness = m_pDS->fv("Sharpness").get_asFloat();
+      settings.m_CustomZoomAmount = m_pDS->fv("ZoomAmount").get_asFloat();
+      settings.m_CustomVerticalShift = m_pDS->fv("VerticalShift").get_asFloat();
+      settings.m_Gamma = m_pDS->fv("Gamma").get_asFloat();
+      settings.m_SubtitleDelay = m_pDS->fv("SubtitleDelay").get_asFloat();
+      settings.m_SubtitleOn = m_pDS->fv("SubtitlesOn").get_asBool();
+      settings.m_SubtitleStream = m_pDS->fv("SubtitleStream").get_asInt();
+      settings.m_ViewMode = m_pDS->fv("ViewMode").get_asInt();
+      settings.m_ResumeTime = m_pDS->fv("ResumeTime").get_asInt();
+      settings.m_InterlaceMethod = (EINTERLACEMETHOD)m_pDS->fv("Deinterlace").get_asInt();
+      settings.m_VolumeAmplification = m_pDS->fv("VolumeAmplification").get_asFloat();
+      settings.m_ScalingMethod = (ESCALINGMETHOD)m_pDS->fv("ScalingMethod").get_asInt();
+      settings.m_StereoMode = m_pDS->fv("StereoMode").get_asInt();
+      settings.m_StereoInvert = m_pDS->fv("StereoInvert").get_asBool();
+      settings.m_SubtitleCached = false;
+      settings.m_VideoStream = m_pDS->fv("VideoStream").get_asInt();
+      settings.m_ToneMapMethod = m_pDS->fv("TonemapMethod").get_asInt();
+      settings.m_ToneMapParam = m_pDS->fv("TonemapParam").get_asFloat();
+      settings.m_Orientation = m_pDS->fv("Orientation").get_asInt();
+      settings.m_CenterMixLevel = m_pDS->fv("CenterMixLevel").get_asInt();
+      m_pDS->close();
+
+      if (settings.m_ToneMapParam == 0.0)
+      {
+        settings.m_ToneMapMethod = VS_TONEMAPMETHOD_REINHARD;
+        settings.m_ToneMapParam = 1.0;
+      }
+      return true;
+    }
+    m_pDS->close();
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
+  }
+  return false;
+}
+
+void CVideoDatabase::SetVideoSettings(const CFileItem &item, const CVideoSettings &settings)
+{
+  int idFile = AddFile(item);
+  SetVideoSettings(idFile, settings);
+}
+
+/// \brief Sets the settings for a particular video file
+void CVideoDatabase::SetVideoSettings(int idFile, const CVideoSettings &setting)
+{
+  try
+  {
+    if (NULL == m_pDB.get())
+      return;
+    if (NULL == m_pDS.get())
+      return;
+    if (idFile < 0)
+      return;
+    std::string strSQL = PrepareSQL("select * from settings where idFile=%i", idFile);
+    m_pDS->query( strSQL );
+    if (m_pDS->num_rows() > 0)
+    {
+      m_pDS->close();
+      // update the item
+      strSQL=PrepareSQL("update settings set Deinterlace=%i,ViewMode=%i,ZoomAmount=%f,PixelRatio=%f,VerticalShift=%f,"
+                       "AudioStream=%i,SubtitleStream=%i,SubtitleDelay=%f,SubtitlesOn=%i,Brightness=%f,Contrast=%f,Gamma=%f,"
+                       "VolumeAmplification=%f,AudioDelay=%f,Sharpness=%f,NoiseReduction=%f,NonLinStretch=%i,PostProcess=%i,ScalingMethod=%i,",
+                       setting.m_InterlaceMethod, setting.m_ViewMode, setting.m_CustomZoomAmount, setting.m_CustomPixelRatio, setting.m_CustomVerticalShift,
+                       setting.m_AudioStream, setting.m_SubtitleStream, setting.m_SubtitleDelay, setting.m_SubtitleOn,
+                       setting.m_Brightness, setting.m_Contrast, setting.m_Gamma, setting.m_VolumeAmplification, setting.m_AudioDelay,
+                       setting.m_Sharpness,setting.m_NoiseReduction,setting.m_CustomNonLinStretch,setting.m_PostProcess,setting.m_ScalingMethod);
+      std::string strSQL2;
+
+      strSQL2=PrepareSQL("ResumeTime=%i,StereoMode=%i,StereoInvert=%i,VideoStream=%i,TonemapMethod=%i,TonemapParam=%f where idFile=%i\n",
+                         setting.m_ResumeTime, setting.m_StereoMode, setting.m_StereoInvert, setting.m_VideoStream,
+                         setting.m_ToneMapMethod, setting.m_ToneMapParam, idFile);
+      strSQL += strSQL2;
+      m_pDS->exec(strSQL);
+      return ;
+    }
+    else
+    { // add the items
+      m_pDS->close();
+      strSQL= "INSERT INTO settings (idFile,Deinterlace,ViewMode,ZoomAmount,PixelRatio, VerticalShift, "
+                "AudioStream,SubtitleStream,SubtitleDelay,SubtitlesOn,Brightness,"
+                "Contrast,Gamma,VolumeAmplification,AudioDelay,"
+                "ResumeTime,"
+                "Sharpness,NoiseReduction,NonLinStretch,PostProcess,ScalingMethod,StereoMode,StereoInvert,VideoStream,TonemapMethod,TonemapParam,Orientation,CenterMixLevel) "
+              "VALUES ";
+      strSQL += PrepareSQL("(%i,%i,%i,%f,%f,%f,%i,%i,%f,%i,%f,%f,%f,%f,%f,%i,%f,%f,%i,%i,%i,%i,%i,%i,%i,%f,%i,%i)",
+                           idFile, setting.m_InterlaceMethod, setting.m_ViewMode, setting.m_CustomZoomAmount, setting.m_CustomPixelRatio, setting.m_CustomVerticalShift,
+                           setting.m_AudioStream, setting.m_SubtitleStream, setting.m_SubtitleDelay, setting.m_SubtitleOn, setting.m_Brightness,
+                           setting.m_Contrast, setting.m_Gamma, setting.m_VolumeAmplification, setting.m_AudioDelay,
+                           setting.m_ResumeTime,
+                           setting.m_Sharpness, setting.m_NoiseReduction, setting.m_CustomNonLinStretch, setting.m_PostProcess, setting.m_ScalingMethod,
+                           setting.m_StereoMode, setting.m_StereoInvert, setting.m_VideoStream, setting.m_ToneMapMethod, setting.m_ToneMapParam, setting.m_Orientation,setting.m_CenterMixLevel);
+      m_pDS->exec(strSQL);
+    }
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s (%d) failed", __FUNCTION__, idFile);
+  }
+}
+
+void CVideoDatabase::EraseAllVideoSettings()
+{
+  try
+  {
+    std::string sql = "DELETE FROM settings";
+
+    CLog::Log(LOGINFO, "Deleting all video settings");
+    m_pDS->exec(sql);
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
+  }
+}
+
+void CVideoDatabase::EraseAllVideoSettings(std::string path)
+{
+  std::string itemsToDelete;
+
+  try
+  {
+    std::string sql = PrepareSQL("SELECT files.idFile FROM files WHERE idFile IN (SELECT idFile FROM files INNER JOIN path ON path.idPath = files.idPath AND path.strPath LIKE \"%s%%\")", path.c_str());
+    m_pDS->query(sql);
+    while (!m_pDS->eof())
+    {
+      std::string file = m_pDS->fv("files.idFile").get_asString() + ",";
+      itemsToDelete += file;
+      m_pDS->next();
+    }
+    m_pDS->close();
+
+    if (!itemsToDelete.empty())
+    {
+      itemsToDelete = "(" + StringUtils::TrimRight(itemsToDelete, ",") + ")";
+
+      sql = "DELETE FROM settings WHERE idFile IN " + itemsToDelete;
+      m_pDS->exec(sql);
+    }
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
+  }
+}
+
+/// \brief EraseVideoSettings() Erases the videoSettings table and reconstructs it
+void CVideoDatabase::EraseVideoSettings(const CFileItem &item)
+{
+  int idFile = GetFileId(item);
+  if (idFile < 0)
+    return;
+
+  try
+  {
+    std::string sql = PrepareSQL("DELETE FROM settings WHERE idFile=%i", idFile);
+
+    CLog::Log(LOGINFO, "Deleting settings information for files %s", CURL::GetRedacted(CURL::GetRedacted(item.GetPath()).c_str()));
+    m_pDS->exec(sql);
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
+  }
+}
+
+void CVideoDatabase::RemoveContentForPath(const std::string& strPath, CGUIDialogProgress *progress /* = NULL */)
+{
+  if(URIUtils::IsMultiPath(strPath))
+  {
+    std::vector<std::string> paths;
+    CMultiPathDirectory::GetPaths(strPath, paths);
+
+    for(unsigned i=0;i<paths.size();i++)
+      RemoveContentForPath(paths[i], progress);
+  }
+
+  try
+  {
+    if (NULL == m_pDB.get()) return ;
+    if (NULL == m_pDS.get()) return ;
+
+    if (progress)
+    {
+      progress->SetHeading(CVariant{700});
+      progress->SetLine(0, CVariant{""});
+      progress->SetLine(1, CVariant{313});
+      progress->SetLine(2, CVariant{330});
+      progress->SetPercentage(0);
+      progress->Open();
+      progress->ShowProgressBar(true);
+    }
+    std::vector<std::pair<int, std::string> > paths;
+    GetSubPaths(strPath, paths);
+    int iCurr = 0;
+    for (const auto &i : paths)
+    {
+      bool bMvidsChecked=false;
+      if (progress)
+      {
+        progress->SetPercentage((int)((float)(iCurr++)/paths.size()*100.f));
+        progress->Progress();
+      }
+
+      if (HasTvShowInfo(i.second))
+        DeleteTvShow(i.second);
+      else
+      {
+        std::string strSQL = PrepareSQL("select files.strFilename from files join movie on movie.idFile=files.idFile where files.idPath=%i", i.first);
+        m_pDS2->query(strSQL);
+        if (m_pDS2->eof())
+        {
+          strSQL = PrepareSQL("select files.strFilename from files join musicvideo on musicvideo.idFile=files.idFile where files.idPath=%i", i.first);
+          m_pDS2->query(strSQL);
+          bMvidsChecked = true;
+        }
+        while (!m_pDS2->eof())
+        {
+          std::string strMoviePath;
+          std::string strFileName = m_pDS2->fv("files.strFilename").get_asString();
+          ConstructPath(strMoviePath, i.second, strFileName);
+          if (HasMovieInfo(strMoviePath))
+            DeleteMovie(strMoviePath);
+          if (HasMusicVideoInfo(strMoviePath))
+            DeleteMusicVideo(strMoviePath);
+          m_pDS2->next();
+          if (m_pDS2->eof() && !bMvidsChecked)
+          {
+            strSQL =PrepareSQL("select files.strFilename from files join musicvideo on musicvideo.idFile=files.idFile where files.idPath=%i", i.first);
+            m_pDS2->query(strSQL);
+            bMvidsChecked = true;
+          }
+        }
+        m_pDS2->close();
+        m_pDS2->exec(PrepareSQL("update path set strContent='', strScraper='', strHash='',strSettings='',useFolderNames=0,scanRecursive=0 where idPath=%i", i.first));
+      }
+    }
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s (%s) failed", __FUNCTION__, strPath.c_str());
+  }
+  if (progress)
+    progress->Close();
+}
+
+void CVideoDatabase::SetScraperForPath(const std::string& filePath, const ScraperPtr& scraper, const VIDEO::SScanSettings& settings)
+{
+  // if we have a multipath, set scraper for all contained paths
+  if(URIUtils::IsMultiPath(filePath))
+  {
+    std::vector<std::string> paths;
+    CMultiPathDirectory::GetPaths(filePath, paths);
+
+    for(unsigned i=0;i<paths.size();i++)
+      SetScraperForPath(paths[i],scraper,settings);
+
+    return;
+  }
+
+  try
+  {
+    if (NULL == m_pDB.get()) return ;
+    if (NULL == m_pDS.get()) return ;
+
+    int idPath = AddPath(filePath);
+    if (idPath < 0)
+      return;
+
+    // Update
+    std::string strSQL;
+    if (settings.exclude)
+    { //NB See note in ::GetScraperForPath about strContent=='none'
+      strSQL=PrepareSQL("update path set strContent='', strScraper='', scanRecursive=0, useFolderNames=0, strSettings='', noUpdate=0 , exclude=1 where idPath=%i", idPath);
+    }
+    else if(!scraper)
+    { // catch clearing content, but not excluding
+      strSQL=PrepareSQL("update path set strContent='', strScraper='', scanRecursive=0, useFolderNames=0, strSettings='', noUpdate=0, exclude=0 where idPath=%i", idPath);
+    }
+    else
+    {
+      std::string content = TranslateContent(scraper->Content());
+      strSQL=PrepareSQL("update path set strContent='%s', strScraper='%s', scanRecursive=%i, useFolderNames=%i, strSettings='%s', noUpdate=%i, exclude=0 where idPath=%i", content.c_str(), scraper->ID().c_str(),settings.recurse,settings.parent_name,scraper->GetPathSettings().c_str(),settings.noupdate, idPath);
+    }
+    m_pDS->exec(strSQL);
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s (%s) failed", __FUNCTION__, filePath.c_str());
+  }
+}
+
+bool CVideoDatabase::ScraperInUse(const std::string &scraperID) const
+{
+  try
+  {
+    if (NULL == m_pDB.get()) return false;
+    if (NULL == m_pDS.get()) return false;
+
+    std::string sql = PrepareSQL("select count(1) from path where strScraper='%s'", scraperID.c_str());
+    if (!m_pDS->query(sql) || m_pDS->num_rows() == 0)
+      return false;
+    bool found = m_pDS->fv(0).get_asInt() > 0;
+    m_pDS->close();
+    return found;
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s(%s) failed", __FUNCTION__, scraperID.c_str());
+  }
+  return false;
+}
+
+ScraperPtr CVideoDatabase::GetScraperForPath( const std::string& strPath )
+{
+  SScanSettings settings;
+  return GetScraperForPath(strPath, settings);
+}
+
+ScraperPtr CVideoDatabase::GetScraperForPath(const std::string& strPath, SScanSettings& settings)
+{
+  bool dummy;
+  return GetScraperForPath(strPath, settings, dummy);
+}
+
+ScraperPtr CVideoDatabase::GetScraperForPath(const std::string& strPath, SScanSettings& settings, bool& foundDirectly)
+{
+  foundDirectly = false;
+  try
+  {
+    if (strPath.empty() || !m_pDB.get() || !m_pDS.get()) return ScraperPtr();
+
+    ScraperPtr scraper;
+    std::string strPath2;
+
+    if (URIUtils::IsMultiPath(strPath))
+      strPath2 = CMultiPathDirectory::GetFirstPath(strPath);
+    else
+      strPath2 = strPath;
+
+    std::string strPath1 = URIUtils::GetDirectory(strPath2);
+    int idPath = GetPathId(strPath1);
+
+    if (idPath > -1)
+    {
+      std::string strSQL=PrepareSQL("select path.strContent,path.strScraper,path.scanRecursive,path.useFolderNames,path.strSettings,path.noUpdate,path.exclude from path where path.idPath=%i",idPath);
+      m_pDS->query( strSQL );
+    }
+
+    int iFound = 1;
+    CONTENT_TYPE content = CONTENT_NONE;
+    if (!m_pDS->eof())
+    { // path is stored in db
+
+      if (m_pDS->fv("path.exclude").get_asBool())
+      {
+        settings.exclude = true;
+        m_pDS->close();
+        return ScraperPtr();
+      }
+      settings.exclude = false;
+
+      // try and ascertain scraper for this path
+      std::string strcontent = m_pDS->fv("path.strContent").get_asString();
+      StringUtils::ToLower(strcontent);
+      content = TranslateContent(strcontent);
+
+      //FIXME paths stored should not have empty strContent
+      //assert(content != CONTENT_NONE);
+      std::string scraperID = m_pDS->fv("path.strScraper").get_asString();
+
+      AddonPtr addon;
+      if (!scraperID.empty() && CServiceBroker::GetAddonMgr().GetAddon(scraperID, addon))
+      {
+        scraper = std::dynamic_pointer_cast<CScraper>(addon);
+        if (!scraper)
+          return ScraperPtr();
+
+        // store this path's content & settings
+        scraper->SetPathSettings(content, m_pDS->fv("path.strSettings").get_asString());
+        settings.parent_name = m_pDS->fv("path.useFolderNames").get_asBool();
+        settings.recurse = m_pDS->fv("path.scanRecursive").get_asInt();
+        settings.noupdate = m_pDS->fv("path.noUpdate").get_asBool();
+      }
+    }
+
+    if (content == CONTENT_NONE)
+    { // this path is not saved in db
+      // we must drill up until a scraper is configured
+      std::string strParent;
+      while (URIUtils::GetParentPath(strPath1, strParent))
+      {
+        iFound++;
+
+        std::string strSQL=PrepareSQL("select path.strContent,path.strScraper,path.scanRecursive,path.useFolderNames,path.strSettings,path.noUpdate, path.exclude from path where strPath='%s'",strParent.c_str());
+        m_pDS->query(strSQL);
+
+        CONTENT_TYPE content = CONTENT_NONE;
+        if (!m_pDS->eof())
+        {
+
+          std::string strcontent = m_pDS->fv("path.strContent").get_asString();
+          StringUtils::ToLower(strcontent);
+          if (m_pDS->fv("path.exclude").get_asBool())
+          {
+            settings.exclude = true;
+            scraper.reset();
+            m_pDS->close();
+            break;
+          }
+
+          content = TranslateContent(strcontent);
+
+          AddonPtr addon;
+          if (content != CONTENT_NONE &&
+              CServiceBroker::GetAddonMgr().GetAddon(m_pDS->fv("path.strScraper").get_asString(), addon))
+          {
+            scraper = std::dynamic_pointer_cast<CScraper>(addon);
+            scraper->SetPathSettings(content, m_pDS->fv("path.strSettings").get_asString());
+            settings.parent_name = m_pDS->fv("path.useFolderNames").get_asBool();
+            settings.recurse = m_pDS->fv("path.scanRecursive").get_asInt();
+            settings.noupdate = m_pDS->fv("path.noUpdate").get_asBool();
+            settings.exclude = false;
+            break;
+          }
+        }
+        strPath1 = strParent;
+      }
+    }
+    m_pDS->close();
+
+    if (!scraper || scraper->Content() == CONTENT_NONE)
+      return ScraperPtr();
+
+    if (scraper->Content() == CONTENT_TVSHOWS)
+    {
+      settings.recurse = 0;
+      if(settings.parent_name) // single show
+      {
+        settings.parent_name_root = settings.parent_name = (iFound == 1);
+      }
+      else // show root
+      {
+        settings.parent_name_root = settings.parent_name = (iFound == 2);
+      }
+    }
+    else if (scraper->Content() == CONTENT_MOVIES)
+    {
+      settings.recurse = settings.recurse - (iFound-1);
+      settings.parent_name_root = settings.parent_name && (!settings.recurse || iFound > 1);
+    }
+    else if (scraper->Content() == CONTENT_MUSICVIDEOS)
+    {
+      settings.recurse = settings.recurse - (iFound-1);
+      settings.parent_name_root = settings.parent_name && (!settings.recurse || iFound > 1);
+    }
+    else
+    {
+      iFound = 0;
+      return ScraperPtr();
+    }
+    foundDirectly = (iFound == 1);
+    return scraper;
+  }
+  catch (...)
+  {
+    CLog::Log(LOGERROR, "%s failed", __FUNCTION__);
+  }
+  return ScraperPtr();
+}
+
+std::string CVideoDatabase::GetContentForPath(const std::string& strPath)
+{
+  SScanSettings settings;
+  bool foundDirectly = false;
+  ScraperPtr scraper = GetScraperForPath(strPath, settings, foundDirectly);
+  if (scraper)
+  {
+    if (scraper->Content() == CONTENT_TVSHOWS)
+    {
+      // check for episodes or seasons.  Assumptions are:
+      // 1. if episodes are in the path then we're in episodes.
+      // 2. if no episodes are found, and content was set directly on this path, then we're in shows.
+      // 3. if no episodes are found, and content was not set directly on this path, we're in seasons (assumes tvshows/seasons/episodes)
+      std::string sql = "SELECT COUNT(*) FROM episode_view ";
+
+      if (foundDirectly)
+        sql += PrepareSQL("WHERE strPath = '%s'", strPath.c_str());
+      else
+        sql += PrepareSQL("WHERE strPath LIKE '%s%%'", strPath.c_str());
+
+      m_pDS->query( sql );
+      if (m_pDS->num_rows() && m_pDS->fv(0).get_asInt() > 0)
+        return "episodes";
+      return foundDirectly ? "tvshows" : "seasons";
+    }
+    return TranslateContent(scraper->Content());
+  }
+  return "";
+}
+
